backend/requirements.txt
asgiref==3.9.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
dj-rest-auth==7.0.1
Django==5.2.6
django-allauth==65.11.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
idna==3.10
pillow==11.3.0
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
requests==2.31.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
uv==0.8.22
